Move cache clearing to app start, out of MyObservations
